// src/app/core/booking/model/bookingFormDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BranchTimeRangeDTO } from './branchTimeRangeDTO';

export interface BookingFormDto { 
    formCode?: string;
    formName?: string;
    formType?: BookingFormDto.FormTypeEnum;
    redirectPage?: string;
    description?: string;
    branchTimeRanges?: Array<BranchTimeRangeDTO>;
    defaultForm?: boolean;
}
export namespace BookingFormDto {
    export type FormTypeEnum = 'BOOKING_FORM' | 'REDIRECT_FORM' | 'EVENT';
    export const FormTypeEnum = {
        BOOKINGFORM: 'BOOKING_FORM' as FormTypeEnum,
        REDIRECTFORM: 'REDIRECT_FORM' as FormTypeEnum,
        EVENT: 'EVENT' as FormTypeEnum
    };
}